StgtoDev

Approved-by: Shraddha Modi

+++ src/components/NotifyMePopup/index.tsx
@@ -36,6 +36,7 @@ import CookieService from "../../services/cookie";
 import { GA_CALLS } from "constants/cookieConsent";
 import { displayPriceWithCommas } from "utils/utility";
 import { isArray } from "lodash";
+import { useHistory } from "react-router";
 
 type Props = {
   basketLineId?: ProductID;
@@ -79,6 +80,7 @@ const NotifyMePopup: React.FC<Props> = ({
   collections
 }) => {
   const { dispatch } = useStore();
+  const history = useHistory();
 
   const user = useContext(UserContext);
   const { closeModal } = useContext(ModalContext);
@@ -193,7 +195,7 @@ const NotifyMePopup: React.FC<Props> = ({
     const size = selectedSize?.size || "";
     // const arr = category?.split(">");
     // const l1 = arr?.[arr.length - 3];
-    const category3 = (sliderImages || [])?.filter(ele => ele?.icon).length
+    const category5 = (sliderImages || [])?.filter(ele => ele?.icon).length
       ? "3d"
       : "non 3d";
     const view3dValue = sliderImages.filter(ele => ele?.icon).length
@@ -251,24 +253,29 @@ const NotifyMePopup: React.FC<Props> = ({
               affiliation: title, // Pass the product name
               coupon: "NA", // Pass the coupon if available
               currency: currency, // Pass the currency code
-              discount: selectedSize?.discountedPriceRecords[currency] || "NA", // Pass the discount amount
+              discount:
+                isSale && selectedSize?.discountedPriceRecords[currency]
+                  ? selectedSize?.discountedPriceRecords[currency]
+                  : "NA", // Pass the discount amount
               index: "NA",
               item_brand: "Goodearth",
-              item_category: category?.split(">")?.join("|"),
-              item_category2: selectedSize?.size, //pass the item category2 ex.Size
-              item_category3: category3, //pass the product type 3d or non 3d
+              item_category: category?.split("/")?.[0],
+              item_category2: category?.split("/")?.[1],
+              item_category3: category?.split("/")?.[2],
+              item_category4: "NA",
+              item_category5: category5,
               item_list_id: "NA", //pass the item list id
               item_list_name: search ? search : "NA", //pass the item list name ex.search results
               item_variant: selectedSize?.size || "",
               // item_category4: l1, //pass the L1,
-              item_category4: "NA",
               // item_category5: collection,
               price: selectedSize?.priceRecords[currency],
               quantity: quantity,
               // dimension12: selectedSize?.color,
               collection_category: isArray(collections)
                 ? collections?.join("|")
-                : collections
+                : collections,
+              price_range: "NA"
             }
           ]
         }
@@ -278,14 +285,16 @@ const NotifyMePopup: React.FC<Props> = ({
 
   const addToBasket = async () => {
     if (selectedSize) {
-      WishlistService.removeFromWishlist(
-        dispatch,
-        selectedSize.id,
-        undefined,
-        sortBy,
-        selectedSize.size
-      );
-      WishlistService.countWishlist(dispatch);
+      if (!history.location.pathname.includes("shared-wishlist")) {
+        WishlistService.removeFromWishlist(
+          dispatch,
+          selectedSize.id,
+          undefined,
+          sortBy,
+          selectedSize.size
+        );
+        // WishlistService.countWishlist(dispatch);
+      }
       setShowLoader(true);
       BasketService.addToBasket(dispatch, selectedSize.id, quantity)
         .then(() => {
@@ -439,9 +448,7 @@ const NotifyMePopup: React.FC<Props> = ({
                     : displayPriceWithCommas(price, currency)}
                 </span>
               ) : (
-                <span
-                  className={badgeType == "B_flat" ? globalStyles.gold : ""}
-                >
+                <span className={badgeType == "B_flat" ? styles.flatPrice : ""}>
                   {selectedSize
                     ? displayPriceWithCommas(
                         selectedSize.priceRecords[currency],
